Bug fixes
Royston Bug 1: No scroll bar when the number of debtors is more than
can be shown on screen.
Steps:
1. Add someone x 5
2. Scroll up and down

Note: There is a scroll bar if there are more receipts than can be
shown on screen

Royston Bug 2: App crashes when you add multiple debtors and then try
to remove them all.
Steps:
1. Add someone x 5
2. Delete them one at a time.
Observed Result:
Each person is deleted until the last one which after deleting will
reappear shortly before the app crashes.
Note: App will crash, and all “deleted” debtors will still be there.

Royston Bug 3: Add a debtor and click into them. Sleep the device. Wake
the device and the back arrow on the top left will be missing.

+++ src/com/nakhbari/calliteven/EntryListItem.java
@@ -5,12 +5,12 @@ import java.util.Calendar;
 public class EntryListItem {
 
 	private String title;
-	private long price;
+	private Double price;
 	private Calendar calendar;
 
 	public EntryListItem() {
 		this.title = "default";
-		this.price = 0;
+		this.price = 0.0;
 		this.calendar = Calendar.getInstance();
 	}
 
@@ -18,7 +18,7 @@ public class EntryListItem {
 		return title;
 	}
 
-	public long getPrice() {
+	public Double getPrice() {
 		return price;
 	}
 
@@ -30,7 +30,7 @@ public class EntryListItem {
 		this.title = name;
 	}
 
-	public void setPrice(long price) {
+	public void setPrice(Double price) {
 		this.price = price;
 	}
